api2/requirements.txt
atomicwrites==1.4.0
attrs==23.2.0
av==11.0.0
dataclasses-json==0.6.7
Deprecated==1.2.15
e2b==0.17.2a62
e2b-code-interpreter==0.0.11a37
flask==3.0.3
googletrans==4.0.2
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.27.1
ipykernel==6.29.5
jiter==0.8.2
jsonpointer==2.1
langchain==0.3.23
langchain-community==0.3.19
langchain-core==0.3.51
langchain-text-splitters==0.3.8
langchain_mongodb==0.6.0
langgraph==0.3.5
langgraph-checkpoint==2.0.18
langgraph-prebuilt==0.1.2
langgraph-sdk==0.1.55
langchain-openai >=0.1.0
libcst==1.6.0
marshmallow==3.26.1
matplotlib==3.9.4
msgpack==1.1.0
nbclient==0.10.2
openai==1.60.0
opencv-python==4.11.0.86
opentelemetry-api==1.29.0
orjson==3.10.15
pillow_heif==0.16.0
protobuf==3.20.3
pyasn1-modules==0.2.8
pydantic==2.7.4
pydantic_core==2.18.4
pydub
pyls-spyder==0.4.0
pytube==15.0.0
pymongo==4.6.1
python-dotenv==1.0.0
PyPDF2>=3.0.1
torch>=1.0.0
quart==0.19.4
quart-cors==0.6.0
scikit-learn==1.6.1
setuptools==69.5.1
SpeechRecognition
streamlit==1.41.1
tavily-python==0.5.1
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.0
transformers>=4.0.0
typing-inspect==0.9.0
typing_extensions==4.12.2
vision-agent==0.2.233
wheel==0.43.0
werkzeug>=3.0.0
zstandard==0.23.0
tf-keras



